requirements.txt
aiofiles==0.8.0
alembic==1.7.7
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
fastapi-pagination==0.9.3
fastapi_amis_admin==0.1.5
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.4
orjson==3.6.8
pandas==1.4.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
python-consul==1.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
sqlalchemy2-stubs==0.0.2a23
sqlmodel==0.0.6
starlette==0.19.1
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
zipp==3.8.0
accelerate==0.22.0
appdirs==1.4.4
loralib==0.1.2
bitsandbytes==0.39.1
black==23.9.1
black[jupyter]
datasets==2.14.5
fire==0.5.0
transformers==4.33.2
sentencepiece==0.1.99
py7zr==0.20.6
scipy==1.11.2
pymilvus==2.3.0
scikit-learn==1.3.0
